requirements.txt
google-generativeai==0.7.2
streamlit==1.36.0
vertexai==1.65.0
